Move slave eeprom methods into own module
